Clean-up environment variables to avoid spoiling tests
Some environment variables are used for setting configurations parameter for
the bonsai API.
This can be a problem if one wants to run tests on their local machine and
they forget to unset these variables.
In particular, some tests expect variables to be empty, and fail if they
aren't.
With this commit tests automatically clean the environment making them
independent from the local environment state.