Introduce BonsaiConnector class for easier sim integration
This commit introduces the BonsaiConnector class which makes it easier for
developers to integrate their sim with the BonsaiPlatform.
The class implements common functions for the communication of a sim with
Bonsai, such as registering the sim and running the event loop for advancing
the simulation state.
The requirements for developers are to create a SimulatorModel class with
some methods implemented.
Then the develeper has to instantiate a BonsaiConnector class and start
the event loop. The connector takes care of registering, running iterations
and episodes, and deregistering the simulation when the job is done.